+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Enhancements:
 
 * Redesign of the admin pages.
 
+Bugs we fixed:
+* Fix redirecting to Progress Planner Dashboard after login.
+
 = 1.7.2 =
 
 * Backported some features from the Progress Planner Pro plugin.